requirements.txt
langchain==0.1.16
faiss-cpu==1.7.4
pydantic==2.6.4
fastapi==0.110.0
uvicorn==0.29.0
spacy==3.7.4
PyMuPDF==1.24.1
python-multipart==0.0.9
pytest==8.1.1
pytest-asyncio==0.23.6
requests==2.31.0
python-dotenv==1.0.1
openai==1.14.3
anthropic==0.21.3
numpy==1.24.3
pandas==2.2.2
pyyaml==6.0.1